backend/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
sqlalchemy
sqlalchemy[asyncio]
pydantic
PyYAML
requests
httpx
aiohttp
h2
redis
apscheduler
kiteconnect
nsepython>=2.97
nsetools>=1.0.10
mibian>=0.1.3
python-jose
passlib[bcrypt]
bcrypt==4.0.1
alembic
aiosqlite
asyncpg
psycopg[binary]
pytest-cov
pytest-asyncio
openpyxl
reportlab
empyrical
pyarrow
hmmlearn
optuna
statsmodels
xgboost
polygon-api-client
websockets>=12.0
python-multipart
